src/requirements.txt
paho-mqtt==1.5.0
redis